requirements.txt
streamlit
truedata_ws==5.0.11
stockstats==0.5.1
gunicorn==23.0.0
wfastcgi==3.0.0
Flask==3.0.3
mysql-connector-python==8.0.21
requests
sqlalchemy